# docs/api/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    TaxonomyCreate:
      description: 'Schema for creating a new taxonomy.


        Used for API requests when creating a new taxonomy.

        Does not include ID or timestamps.'
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional taxonomy description
          title: Description
        identifier:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          description: Taxonomy CURIE
          title: Identifier
        skos_uri:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional SKOS URI for alignment
          title: Skos Uri
        title:
          description: Taxonomy title
          maxLength: 255
          minLength: 1
          title: Title
          type: string
      required:
      - title
      title: TaxonomyCreate
      type: object
    TaxonomyListResponse:
      description: Schema for paginated taxonomy list responses.
      properties:
        items:
          description: List of taxonomies
          items:
            $ref: '#/components/schemas/TaxonomyResponse'
          title: Items
          type: array
        limit:
          description: Items per page
          minimum: 1.0
          title: Limit
          type: integer
        offset:
          description: Offset from start
          minimum: 0.0
          title: Offset
          type: integer
        total:
          description: Total count of taxonomies
          minimum: 0.0
          title: Total
          type: integer
      required:
      - items
      - total
      - limit
      - offset
      title: TaxonomyListResponse
      type: object
    TaxonomyResponse:
      description: 'Schema for taxonomy responses.


        Used for API responses. Includes all fields including ID and timestamps.'
      properties:
        created_at:
          description: Creation timestamp
          format: date-time
          title: Created At
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional taxonomy description
          title: Description
        id:
          description: Unique identifier
          format: uuid
          title: Id
          type: string
        identifier:
          description: Taxonomy CURIE
          maxLength: 255
          minLength: 1
          title: Identifier
          type: string
        skos_uri:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional SKOS URI for alignment
          title: Skos Uri
        title:
          description: Taxonomy title
          maxLength: 255
          minLength: 1
          title: Title
          type: string
        updated_at:
          description: Last update timestamp
          format: date-time
          title: Updated At
          type: string
      required:
      - title
      - identifier
      - id
      - created_at
      - updated_at
      title: TaxonomyResponse
      type: object
...
    TaxonomyUpdate:
      description: 'Schema for updating an existing taxonomy.


        All fields are optional to support partial updates.'
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        skos_uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Skos Uri
        title:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          title: Title
      title: TaxonomyUpdate
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact:
    email: mike@mikewaters.net
    name: LifeOS Development Team
    url: https://github.com/LifeOS/Substrate
  description: REST API for topic taxonomy and knowledge management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Ontology API
  version: 0.1.0
  x-api-id: ontology-api
  x-audience: internal
openapi: 3.1.0
paths:
...
  /taxonomies:
    get:
      description: "List taxonomies with pagination.\n\nArgs:\n    service: Taxonomy service dependency\n    limit: Maximum\
        \ number of results (1-100)\n    offset: Number of results to skip\n\nReturns:\n    Paginated list of taxonomies"
      operationId: list_taxonomies_taxonomies_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyListResponse'
              example:
                items:
                - &id001
                  id: 00000000-0000-0000-d315-cb6cbdc35142
                  identifier: tx:tech
                  title: Technology
                  description: Topics related to technology domains
                  skos_uri: null
                  created_at: '2025-11-04T02:55:11.302186Z'
                  updated_at: '2025-11-04T02:55:11.302186Z'
                - id: 00000000-0000-0000-188d-6696711aee7d
                  identifier: tx:health
                  title: Health & Wellness
                  description: Physical and mental health topics
                  skos_uri: null
                  created_at: '2025-11-04T02:55:11.302186Z'
                  updated_at: '2025-11-04T02:55:11.302186Z'
                total: 2
                limit: 50
                offset: 0
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List taxonomies
      tags:
      - taxonomies
    post:
      description: "Create a new taxonomy.\n\nArgs:\n    taxonomy: Taxonomy creation data\n    service: Taxonomy service dependency\n\
        \nReturns:\n    Created taxonomy\n\nRaises:\n    ValidationError: If validation fails"
      operationId: create_taxonomy_taxonomies_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaxonomyCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a new taxonomy
      tags:
      - taxonomies
  /taxonomies/{taxonomy_id}:
    delete:
      description: "Delete a taxonomy.\n\nArgs:\n    taxonomy_id: Taxonomy ID\n    service: Taxonomy service dependency\n\
        \    cascade: If True, delete associated topics\n\nRaises:\n    NotFoundError: If taxonomy not found"
      operationId: delete_taxonomy_taxonomies__taxonomy_id__delete
      parameters:
      - description: Taxonomy ID
        in: path
        name: taxonomy_id
        required: true
        schema:
          description: Taxonomy ID
          format: uuid
          title: Taxonomy Id
          type: string
      - description: Delete associated topics
        in: query
        name: cascade
        required: false
        schema:
          default: true
          description: Delete associated topics
          title: Cascade
          type: boolean
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete a taxonomy
      tags:
      - taxonomies
    get:
      description: "Get a taxonomy by ID.\n\nArgs:\n    taxonomy_id: Taxonomy ID\n    service: Taxonomy service dependency\n\
        \nReturns:\n    Taxonomy\n\nRaises:\n    NotFoundError: If taxonomy not found"
      operationId: get_taxonomy_taxonomies__taxonomy_id__get
      parameters:
      - description: Taxonomy ID
        in: path
        name: taxonomy_id
        required: true
        schema:
          description: Taxonomy ID
          format: uuid
          title: Taxonomy Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyResponse'
              example: *id001
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a taxonomy by ID
      tags:
      - taxonomies
    put:
      description: "Update a taxonomy.\n\nArgs:\n    taxonomy_id: Taxonomy ID\n    updates: Fields to update\n    service:\
        \ Taxonomy service dependency\n\nReturns:\n    Updated taxonomy\n\nRaises:\n    NotFoundError: If taxonomy not found"
      operationId: update_taxonomy_taxonomies__taxonomy_id__put
      parameters:
      - description: Taxonomy ID
        in: path
        name: taxonomy_id
        required: true
        schema:
          description: Taxonomy ID
          format: uuid
          title: Taxonomy Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaxonomyUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update a taxonomy
      tags:
      - taxonomies
...
servers:
- description: Development server
  url: http://localhost:8000
- description: Production server
  url: https://api.lifeos.dev